bootstrap only installs some things for osx

+++ bootstrap.sh
@@ -6,7 +6,7 @@ cd "$(dirname "${BASH_SOURCE}")";
 
 git pull origin master;
 
-function doIt() {
+function sync() {
     rsync --exclude ".git/" \
         --exclude ".DS_Store" \
         --exclude "bootstrap.sh" \
@@ -17,21 +17,6 @@ function doIt() {
         -av --no-perms . ~
     source ~/.bash_profile;
 }
-
-if [ "$1" == "--force" -o "$1" == "-f" ]; then
-    doIt;
-else
-    read -p "This may overwrite existing files in your home directory. Are you sure? (y/n) " -n 1;
-    echo "";
-    if [[ $REPLY =~ ^[Yy]$ ]]; then
-        doIt;
-    fi;
-fi;
-unset doIt;
-
-# Install XCcode command line tools
-echo "Installing XCode command line tools..."
-xcode-select --install
 
 # Homebrew OS X package manager
 function install_homebrew() {
@@ -45,39 +30,6 @@ function install_homebrew() {
     fi
 }
 
-# call homebrew and homebrew cask scripts (installs NPM, etc)
-read -p "Install Homebrew and all packages (y/n)? " choice
-case "$choice" in
-  y|Y ) install_homebrew;;
-  n|N ) echo "Skipping homebrew";;
-  * ) echo "invalid answer";;
-esac
-
-# Install Node.js (Latest 'Stable')
-mkdir -p ~/.nvm
-# Setup NVM
-export NVM_DIR=~/.nvm
-. $(brew --prefix nvm)/nvm.sh
-[[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm"
-echo "Installing Node.js (Latest 'stable')..."
-nvm install node # "node" is an alias for latest stable
-nvm alias default node # set "node" as the default
-
-# update / install npm packages
-# Check for npm
-if [ `type -P npm` ]; then
-    # Installing NPM packages...
-    echo "Installing NPM packages..."
-    npm install node-inspector --global --quiet
-    npm install bower --global --quiet
-    [[ $? ]] && echo "Done"
-else
-    printf "\n"
-    echo "Error: npm not found."
-    printf "Aborting... try installing node packages manually\n"
-    exit
-fi
-
 # Ruby version manager
 function install_rvm() {
     curl -sSL https://get.rvm.io | bash
@@ -88,24 +40,80 @@ function install_rvm() {
     rvm install ruby-head
 }
 
-read -p "Install rvm with curl | bash, Continue (y/n)? " choice
-case "$choice" in
-  y|Y ) install_rvm;;
-  n|N ) echo "Skipping rvm";;
-  * ) echo "invalid answer";;
-esac
+if [ "$1" == "--force" -o "$1" == "-f" ]; then
+    sync;
+else
+    read -p "This may overwrite existing files in your home directory. Are you sure? (y/n) " -n 1;
+    echo "";
+    if [[ $REPLY =~ ^[Yy]$ ]]; then
+        sync;
+    fi;
+fi;
 
-# Install Python (Latest 'Stable')
-echo "Installing Python (Latest 'stable')..."
-pyenv install `pyenv install --list | grep -v - | grep -v b | tail -1`
 
-# Setup OS X system defaults
-read -p "Setup OS X system defaults (y/n)? " choice
-case "$choice" in
-  y|Y ) source defaults.sh;;
-  n|N ) echo "Skipping OS X defaults";;
-  * ) echo "invalid answer";;
-esac
+if [[ $OSTYPE == darwin* ]]; then
+
+    # Install XCcode command line tools
+    echo "Installing XCode command line tools..."
+    xcode-select --install
+
+    # call homebrew and homebrew cask scripts (installs NPM, etc)
+    read -p "Install Homebrew and all packages (y/n)? " choice
+    case "$choice" in
+      y|Y ) install_homebrew;;
+      n|N ) echo "Skipping homebrew";;
+      * ) echo "invalid answer";;
+    esac
+
+    # Install Node.js (Latest 'Stable')
+    mkdir -p ~/.nvm
+    # Setup NVM
+    export NVM_DIR=~/.nvm
+    . $(brew --prefix nvm)/nvm.sh
+    [[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm"
+    echo "Installing Node.js (Latest 'stable')..."
+    nvm install node # "node" is an alias for latest stable
+    nvm alias default node # set "node" as the default
+
+    # update / install npm packages
+    # Check for npm
+    if [ `type -P npm` ]; then
+        # Installing NPM packages...
+        echo "Installing NPM packages..."
+        npm install node-inspector --global --quiet
+        npm install bower --global --quiet
+        [[ $? ]] && echo "Done"
+    else
+        printf "\n"
+        echo "Error: npm not found."
+        printf "Aborting... try installing node packages manually\n"
+        exit
+    fi;
+
+    read -p "Install rvm with curl | bash, Continue (y/n)? " choice
+    case "$choice" in
+      y|Y ) install_rvm;;
+      n|N ) echo "Skipping rvm";;
+      * ) echo "invalid answer";;
+    esac
+
+    # Install Python (Latest 'Stable')
+    echo "Installing Python (Latest 'stable')..."
+    pyenv install `pyenv install --list | grep -v - | grep -v b | tail -1`
+
+    # Setup OS X system defaults
+    read -p "Setup OS X system defaults (y/n)? " choice
+    case "$choice" in
+      y|Y ) source defaults.sh;;
+      n|N ) echo "Skipping OS X defaults";;
+      * ) echo "invalid answer";;
+    esac
+fi;
+
+# cleanup
+unset sync;
+unset install_homebrew;
+unset install_rvm;
 
 # finish up
 echo
